[iOS] Get initial vehicle status from backend (#6494)
### Goals
Get initial driver vehicle status from backend instead of relying on status from client's local 
storage.

Previously, whether or not a vehicle was registered was only represented locally on the client 
using user defaults storage, and a vehicle's initial status was assumed to be "offline" if the 
vehicle was registered, which had a couple issues:
- The registration state could become inaccurate if the user changed environments or deleted 
and re-installed the app.
- If the app was restarted while the driver was "online", the initial state on startup would 
inaccurately reflect being "offline."

This PR modifies the iOS driver app to check the backend for a vehicle's status on startup, 
which enables us to more accurately determine if a driver's vehicle has been registered yet, and 
whether or not the driver is currently "online" or "offline" when the app starts.

Also, this PR modifies `StateMachine` to take in a `logger` to use as a `init` parameter with a 
default value. This enables tests that exercise logic relying on `StateMachine` (and potentially 
generate logged errors) to use `ConsoleLogger` instead of always relying on the logger 
provided by `LoggerDependencyRegistry`, which is generally not initialized for tests.

### Implementation
Add `VehicleStatus` model, which represents a driver's current vehicle status. Currently, this 
can be one of `ready`, `notReady`, or `unregistered`.

Add new function to `[Default]DriverVehicleInteractor`, `getVehicleStatus(vehicleId: String) -> 
Single<VehicleStatus>`, which is responsible for getting a vehicle's state from the backend and 
translating it to an appropriate `VehicleStatus`.

Modify `MainViewState` model to include an additional `unknown` state, which represents the 
situation where the client hasn't retrieved a `VehicleStatus` from the backend yet and is waiting 
for a response before transitioning to the appropriate driver UX flow.

Modify `DefaultMainViewModel` to start with the initial state of `MainViewState.unknown` until 
the `VehicleStatus` is retrieved from the backend. This replaces the previous behavior where 
`DefaultMainViewModel` would read from local storage and default to `unregistered` or 
`offline` depending on whether or not any registration info was present. 
`DefaultMainViewModel` filters the results provided to observers of `MainViewState`, so that 
they should not receive an update until `MainViewState` is transitions away from being 
`unknown` (i.e., they should only receive updates when the state is confirmed to be one of 
`vehicleUnregistered`, `offline`, or `online`).

Modfy `DefaultMainViewModelTest` so tests reflect changes in behavior and to add tests that 
verify new initial state behavior and behavior when completing vehicle registration.

Modify `StateMachine` to take in a `Logger` as an `init` param with a default value of 
`LoggerDependencyRegistry.instance.logger` instead of always relying on 
`LoggerDependencyRegistry.instance.logger`. This enables `ConsoleLogger` to be used for 
running tests.

Modify a number of places that rely on `StateMachine` to take in their own `Logger` 
parameters so they can also be modified to use a `ConsoleLogger` for tests and so they can 
provide the appropriate type of logger to `StateMachine`.

Update some tests to use `ConsoleLogger` where appropriate.

### Suggested reading order
- `VehicleStatus`
- `[Default]DriverVehicleInteractor`
- `MainViewState` /  `DefaultMainViewModel` / `DefaultMainViewModelTest`
- `StateMachine`
- remaining files in `public/rideos-sdk-ios/RideOsDriver`
- remaining files in `public/rideos-sdk-ios/RideOsRider`
- remaining files in `public/rideos-sdk-ios/RideOsTests`