[public] Turn off poor GPS behavior when running sim (#6575)
### Goals
When running `MapboxNavigationViewController` in the simulator, default to not simulating 
location unless that developer settings is explicitly turned on. This is different from the previous 
behavior, where location would be simulated when Mapbox detected a poor GPS signal.

This is done to reduce developer confusion when they run the driver app in the simulator and 
expect location simulation to explicitly be turned on or off, rather than being active depending 
on when Mapbox detects a poor GPS signal.

### Implementation
Add check `MapboxNavigationViewController` to detect if we are running in the simulator, and, 
if so, make the default simulation behavior `.never` instead of `.poorGPS`. Also, add comment 
explaining the reasoning behind this check.