[public] Add progress UI for offline/idle states (#6601)
### Goals
Modify `OfflineViewController` and `IdleViewController` to make it more obvious when a 
network request is in progress.

Also, indicate when an error has occurred with a network request via an alert with recovery 
instructions for the user.

Finally, modify language used for the offline and idle driver UX to be more user-friendly.

### Implementation
Add new view models, `DefaultOfflineViewModel` and `DefaultIdleViewModel`, which are now 
responsible for marking a vehicle as ready or not and also representing the progress for those 
requests (including an error state, if appropriate). Also, add corresponding tests for new view 
models.

Modify `OfflineViewController` and `IdleViewController` to make use of newly added view
models.

Modify `OfflineDialogView` and `IdleDialogView` to include progress indicators, to provide 
affordances for toggling whether or not the dialog is currently reflecting the "in progress" state,
and to provide affordances for enabling / disable the dialog's action button.

Rename `GoOnlineListener` / `GoOfflineListener` function names to more accurately describe 
the context when they are called, i.e., after the process of going online or offline has completed, 
rather than indicating that those process should start.

Modify `DefaultMainViewModel` to no longer be responsible for marking a vehicle ready or not 
on the backend. This is now handled by `DefaultOfflineViewModel` and `DefaultIdleViewModel`.

Modify `StateMachine` to add function, `getCurrentState`, that returns the current state 
without requiring the use of an `Observable`, which matches the Android `StateMachine`'s 
interface. Also, update name for function that returns `Observable` state to match Android.

Slightly simplify `DefaultDriverVehicleInteractor`'s implementation by replacing use of 
`Observable` and `ignoreElements` with `Completable`.

### Suggested reading order
- `OfflineViewModel` / `DefaultOfflineViewModel` / `DefaultOfflineViewModelTest`
- `OfflineDialogView`
- `OfflineViewController`
- `IdleViewModel` / `DefaultIdleViewModel` / `DefaultIdleViewModelTest`
- `IdleDialogView`
- `IdleViewController`
- `DefaultMainViewModel`
- the rest...

### Future improvements
Perform similar treatment for other portions of driver UX where network requests occur.